// web/src/contracts/planogram.schema.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface Planogram {
  planogram_id: string;
  name: string;
  source: "manual" | "video";
  /**
   * @minItems 1
   */
  bays: [Bay, ...Bay[]];
  /**
   * @minItems 1
   */
  skus: [Sku, ...Sku[]];
  creatives: Creative[];
}
/**
 * This interface was referenced by `Planogram`'s JSON-Schema
 * via the `definition` "bay".
 */
export interface Bay {
  bay_id: string;
  type: "shelf" | "endcap";
  width_m: number;
  height_m: number;
  station: {
    /**
     * @minItems 3
     * @maxItems 3
     */
    camera_pos: [number, number, number];
    /**
     * @minItems 3
     * @maxItems 3
     */
    look_at: [number, number, number];
  };
  /**
   * @minItems 1
   */
  shelves: [Shelf, ...Shelf[]];
  ad_slots: AdSlot[];
}
/**
 * This interface was referenced by `Planogram`'s JSON-Schema
 * via the `definition` "shelf".
 */
export interface Shelf {
  shelf_id: string;
  height_m: number;
  level: "top" | "above_eye" | "eye" | "below_eye" | "bottom";
  slots: Slot[];
}
/**
 * This interface was referenced by `Planogram`'s JSON-Schema
 * via the `definition` "slot".
 */
export interface Slot {
  slot_id: string;
  sku_id: string | null;
  facings: number;
  x_m: number;
  width_m: number;
  height_m: number;
  confidence?: number;
}
/**
 * This interface was referenced by `Planogram`'s JSON-Schema
 * via the `definition` "ad_slot".
 */
export interface AdSlot {
  ad_slot_id: string;
  type: "shelf_talker" | "endcap_header" | "floor_decal" | "screen";
  attached_to: string;
  x_m: number;
  width_m: number;
  creative_id: string | null;
  confidence?: number;
}
/**
 * This interface was referenced by `Planogram`'s JSON-Schema
 * via the `definition` "sku".
 */
export interface Sku {
  sku_id: string;
  name: string;
  brand: string;
  category: string;
  price: number;
  promo: boolean;
  texture_url: string;
  /**
   * @minItems 3
   * @maxItems 3
   */
  color_lab: [number, number, number];
}
/**
 * This interface was referenced by `Planogram`'s JSON-Schema
 * via the `definition` "creative".
 */
export interface Creative {
  creative_id: string;
  brand: string;
  texture_url: string;
}